cocos2d-xライブラリをバージョンアップ

+++ cocos2d/cocos/3d/CCBundleReader.h
@@ -87,7 +87,7 @@ public:
     /**
      * Returns the position of the file pointer.
      */
-    ssize_t tell();
+    long int tell();
 
     /**
      * Sets the position of the file pointer.
